local_backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Vector DB
chromadb==0.4.18

# Embeddings
# sentence-transformers==2.2.2  # Optional - only needed as fallback when Ollama unavailable
# nomic-embed-text # Using Ollama instead

# PDF parsing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Text processing
# tiktoken==0.5.2  # Optional - only for accurate token counting (has fallback approximation)
# spacy==3.7.2  # Optional - not needed when using Ollama

# Auth
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg[binary]==3.1.18

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
